LPS-63587 Explicitly set a JAXWS SPI provider
Somehow, SPI packages are exported twice: one by core bootstrap and
another one by the JDK. If we don't specify the provider explicitly, a
default one will be loaded by the JDK using one version of the
package. The problem is that later it will fail with a
ClassCastException while trying to reference it through an interface
loaded with the other package version.

The choice of CXF is arbitrary: none of its classes are exported by the
core (so less chance of a conflict), and it is supported by OpenCMIS.