LPS-70420 Protect against NPE
The `toFolderOrFileEntry` method in `CMISRepository` may return null,
which later the repo will try to save in the cache. While this seem
not very probable, make sure we'll handle this case -- after all it is
perfectly reasonable (and possible) to connect to a CMIS repo which
contains non folder or file objects.